requirements.txt
coloredlogs==7.3
matplotlib==3.9.2
numpy==2.1.2
PyYAML==6.0.2
PyYAML==5.4.1
PyYAML==6.0.2
torch==2.2.2
tqdm==4.66.5
tqdm==4.57.0
